tests/wpt/tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/document.open-01.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>document.open in XHTML</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com"/>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#opening-the-input-stream"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function() {
  assert_throws_dom("INVALID_STATE_ERR", function() {
    document.open();
  }, "document.open in XHTML should throw an INVALID_STATE_ERR ");
}, "document.open in XHTML");
</script>
</body>
</html>
